Remake route /users/isloggedin
This route originally was a POST HTTP verb, but it has been changed to a
GET, which fits better with its purpose: getting something from the API
under certain circumstances.

Furthermore, the route originally returned everything about the user,
including the hashed password. This has been patched, for security's
sake.